refactor: 전체적인 Refactoring
