+++ js/database.js
@@ -1,4 +1,34 @@
-export const address = ['서울 마포구 월드컵북로 434 상암IT센터 1층 116호', '서울 마포구 월드컵북로58길 10 더팬빌딩 2층', '서울 마포구 월드컵북로56길 9 우리기술사옥 2층','서울 마포구 월드컵북로56길 19 드림타워 2층','서울 마포구 월드컵북로54길 25 B1층 B109호','서울 마포구 월드컵북로54길 25 상암DMC푸르지오시티 지하1층 B107호','서울 마포구 월드컵북로54길 25 지하1층 B108호', '서울특별시 마포구 월드컵북로54길 25', '서울특별시 마포구 월드컵북로54길 17 109호', '서울특별시 마포구 상암동 257', '서울특별시 마포구 월드컵북로56길 19', '서울특별시 마포구 상암동 1597', '서울특별시 마포구 월드컵북로54길 25', '서울특별시 마포구 월드컵북로54길 25 푸르지오시티 218호', '서울 마포구 월드컵북로54길 17 사보이시티 B동 지하1층 B126호', '서울 마포구 월드컵북로54길 17 사보이시티 지하 1층 B-132A 도락', '서울 마포구 월드컵북로54길 17 지하1층 124호 경양식당','서울 마포구 월드컵북로54길 17 사보이시티 DMC 지하 1층', '서울 마포구 월드컵북로54길 17 사보이시티 DMC 지하1층 125호', '서울 마포구 월드컵북로54길 17 1층 110호 장정정', '서울 마포구 월드컵북로54길 17 사보이시티디엠씨 지하1층 b116호 달리181', '서울 마포구 월드컵북로54길 17 사보이시티 B동 2층 206,210호', '서울 마포구 월드컵북로54길 25 상암DMC푸르지오시티 지하1층 B104호, 105호', '서울 마포구 월드컵북로 402 케이지아이티센터2층', '서울 마포구 월드컵북로 400 서울경제진흥원 지하1층', '서울 마포구 월드컵북로 400 한국영상자료원 1층 A111호', '서울 마포구 월드컵북로 396 누리꿈스퀘어 공동제작센터 2층', '서울 마포구 월드컵북로 381', '서울 마포구 성암로 267 MBC몰파크 지하1층 B117호', '서울 마포구 성암로 267 지하1층 비110호', '서울 마포구 상암산로 76 YTN 뉴스퀘어 지하1층', '서울 마포구 월드컵북로 361 1층 103호, 서울 마포구 성암로 229-1', '서울 마포구 월드컵북로 375 이안1단지 1층', '서울 마포구 월드컵북로48길 29-12', '서울 마포구 성암로15길 43 1층', '서울특별시 마포구 월드컵북로 400', '서울특별시 마포구 월드컵북로 402', '서울특별시 마포구 월드컵북로 396', '서울 마포구 상암산로 38 지하 1층', '상암동 1590번지 108호 상암ITTOWER 마포구 서울특별시 KR', '서울 마포구 월드컵북로 434 상암it타워 1층 101호', '서울특별시 마포구 상암동 1597', '서울특별시 마포구 상암동 1596번지 S-CITY 1층 116호', '서울특별시 마포구 상암동 1601', '서울 마포구 월드컵북로 402 kgit센터 109호', 'KR 서울특별시 마포구 상암동 6-2번지 1층 R112호', '서울특별시 마포구 월드컵북로 402', '서울특별시 마포구 월드컵북로54길 17', '서울특별시 마포구 월드컵북로 396', '상암동 1605번지 상암동누리꿈스퀘어 지하1층 1066호 마포구 서울특별시 KR', '서울특별시 마포구 월드컵북로 396 누리꿈스퀘어 지하 1층 1019호', '서울 마포구 월드컵북로 396 누리꿈스퀘어 B1034 스시앤도시락', '서울 마포구 월드컵북로54길 17 1층 108호', '서울 마포구 월드컵북로 396 누리꿈스퀘어 지하1층 1081호 돈까스브로스', '서울 마포구 월드컵북로 396 누리꿈스퀘어 지하1층 B1079, 1080호 포36거리', '서울 마포구 월드컵북로 396 디지털파빌리온 지하1층 1041호', '서울 마포구 월드컵북로 396 누리꿈스퀘어 지하1층 1109-1호',]
-export const representative_food = ['김치찌개', '태국음식(나시고랭, 팟타이 등)', '샤브샤브(칼국수)', '수제버거', '샤브샤브', '일식(규카츠, 덮밥)', '우육면, 탄탄면, 샤오롱바오', '우육면, 샤오롱바오', '곰탕', '돼지불백, 제주흑돼지 등', '차돌 쌀국수', '중식(짜장, 짬뽕 등)', '돈카츠(로스카츠, 히레카츠)', '가츠동', '나시고랭, 팟타이 등', '제주해장국', '경양식 돈까스, 함박스테이크 등', '샤브샤브', '캘리포니아롤', '일식 덮밥(사케동, 스테이크덮밥 등)', '양식(파스타, 빠네) - 데이트 추천', '평양냉면, 갈비탕(+고기집)', '꼬막 음식점', '참치 전문점(강사님 사주세요)', '보쌈정식, 감자옹심이 등', '양식(피자, 파스타 등) - 데이트 추천', '낙지볶음, 낙곱새볶음', '돈까스', '중식(짜장, 짬뽕 등)', '탄탄멘, 스테이크덮밥', '콩나물 국밥', '순대국밥, 나주곰탕 등', '국밥(a.k.a. 부산아지매)', '양평 해장국', '뼈해장국, 삼계탕 등', '햄버거', '햄버거', '햄버거', '구내식당','김밥천국 같은 곳', '김밥집', '김밥집', '김밥천국 같은 곳', '김밥 천국임.', '덮밥 류', '김밥집', '도시락', '쌀국수', '라면', '순두부찌개집(default 비빔밥 가능)', '소,머리국밥', '초밥', '일식류의 덮밥, 카레, 라멘', '돈까스', '쌀국수', '쌀국수', '떡볶이', ];
-export const price_per_person = ['10000', '20000u', '11000', '11700u', '14000', '19000','9500','15000', '12500', '10000', '15000', '10000', '15000','15000', '12000', '11000', '11000', '15000', '10000', '15000', '16000', '17000', '13000', '35000', '13000', '20000u', '14000', '12000', '10000', '10800', '9000', '13000', '11500', '15000', '11000', '7000', '7000', '7000', '6500', '8000', '김밥 한 줄 5000원, 닭강정 세트 10000', '한 줄 5500', '8000', '8000', '8500', '8500', '7000', '10000', '5000', '8500', '9000', '12900원 ~ 14900', '13000', '11900', '9900', '11000', '14000', ];
-export const name = ['김치도가 상암점', '생어거스틴 상암DMC점', '등촌샤브칼국수 상암점', '바스버거 상암DMC점', '청록미나리식당 상암점', '후라토식당 상암직영점','룽탕 상하이딤섬', '샤오바오 상암점', '서당골 상암점', '동남집', '미분당 상암점', '홍콩반점0410 상암DMC점', '은옥', '온돈부리', '온타이키친', '도락', '경양식당', '채선당 상암DMC점', '미미롤집', '장정정 상암정', '달리181 상암점', '배꼽집 본점', '여수낭만식당', '참치공방 상암점', '시골보쌈&감자옹심이 상암점', '뚜띠쿠치나 상암점', '개미집 상암점', '마이클돈까스 본점', '용무있습니까 상암점', '미도인 상암엠비씨', '전주현대옥 상암DMC점', '김둘레순대국 상암2/4호점', '부산아지매국밥 상암DMC점', '상암 양평 해장국', '푸짐한왕뼈감자탕', '맥도날드 서울상암DMC', '롯데리아 상암점', 'KFC 상암DMC점', 'JTBC 구내식당', '최가네김밥', '오토김밥상암점', '청담마녀김밥 상암점', '김가네상암푸르지오', '김밥천국 디엠씨점', '오늘은덮밥', '여우애김밥 상암DMC점', '한솥도시락 상암DMC점', '포메인레드 상암DMC점', '틈새라면', '새벽집', '용길이네소머리국밥', '스시앤도시락 상암점', '핵밥 상암DMC점', '돈까스브로스 상암DMC점', '포36거리 상암', '동남쌀국 상암점', '응급실국물떡볶이 상암점', ];
+import { createNameList } from './util.js';
+import { sortNameList } from './accel/array.js';
+
+// 가게 정보 데이터베이스
+export const ADDRESS = ['서울 마포구 월드컵북로 434 상암IT센터 1층 116호', '서울 마포구 월드컵북로58길 10 더팬빌딩 2층', '서울 마포구 월드컵북로56길 9 우리기술사옥 2층','서울 마포구 월드컵북로56길 19 드림타워 2층','서울 마포구 월드컵북로54길 25 B1층 B109호','서울 마포구 월드컵북로54길 25 상암DMC푸르지오시티 지하1층 B107호','서울 마포구 월드컵북로54길 25 지하1층 B108호', '서울특별시 마포구 월드컵북로54길 25', '서울특별시 마포구 월드컵북로54길 17 109호', '서울특별시 마포구 상암동 257', '서울특별시 마포구 월드컵북로56길 19', '서울특별시 마포구 상암동 1597', '서울특별시 마포구 월드컵북로54길 25', '서울특별시 마포구 월드컵북로54길 25 푸르지오시티 218호', '서울 마포구 월드컵북로54길 17 사보이시티 B동 지하1층 B126호', '서울 마포구 월드컵북로54길 17 사보이시티 지하 1층 B-132A 도락', '서울 마포구 월드컵북로54길 17 지하1층 124호 경양식당','서울 마포구 월드컵북로54길 17 사보이시티 DMC 지하 1층', '서울 마포구 월드컵북로54길 17 사보이시티 DMC 지하1층 125호', '서울 마포구 월드컵북로54길 17 1층 110호 장정정', '서울 마포구 월드컵북로54길 17 사보이시티디엠씨 지하1층 b116호 달리181', '서울 마포구 월드컵북로54길 17 사보이시티 B동 2층 206,210호', '서울 마포구 월드컵북로54길 25 상암DMC푸르지오시티 지하1층 B104호, 105호', '서울 마포구 월드컵북로 402 케이지아이티센터2층', '서울 마포구 월드컵북로 400 서울경제진흥원 지하1층', '서울 마포구 월드컵북로 400 한국영상자료원 1층 A111호', '서울 마포구 월드컵북로 396 누리꿈스퀘어 공동제작센터 2층', '서울 마포구 월드컵북로 381', '서울 마포구 성암로 267 MBC몰파크 지하1층 B117호', '서울 마포구 성암로 267 지하1층 비110호', '서울 마포구 상암산로 76 YTN 뉴스퀘어 지하1층', '서울 마포구 월드컵북로 361 1층 103호, 서울 마포구 성암로 229-1', '서울 마포구 월드컵북로 375 이안1단지 1층', '서울 마포구 월드컵북로48길 29-12', '서울 마포구 성암로15길 43 1층', '서울특별시 마포구 월드컵북로 400', '서울특별시 마포구 월드컵북로 402', '서울특별시 마포구 월드컵북로 396', '서울 마포구 상암산로 38 지하 1층', '상암동 1590번지 108호 상암ITTOWER 마포구 서울특별시 KR', '서울 마포구 월드컵북로 434 상암it타워 1층 101호', '서울특별시 마포구 상암동 1597', '서울특별시 마포구 상암동 1596번지 S-CITY 1층 116호', '서울특별시 마포구 상암동 1601', '서울 마포구 월드컵북로 402 kgit센터 109호', 'KR 서울특별시 마포구 상암동 6-2번지 1층 R112호', '서울특별시 마포구 월드컵북로 402', '서울특별시 마포구 월드컵북로54길 17', '서울특별시 마포구 월드컵북로 396', '상암동 1605번지 상암동누리꿈스퀘어 지하1층 1066호 마포구 서울특별시 KR', '서울특별시 마포구 월드컵북로 396 누리꿈스퀘어 지하 1층 1019호', '서울 마포구 월드컵북로 396 누리꿈스퀘어 B1034 스시앤도시락', '서울 마포구 월드컵북로54길 17 1층 108호', '서울 마포구 월드컵북로 396 누리꿈스퀘어 지하1층 1081호 돈까스브로스', '서울 마포구 월드컵북로 396 누리꿈스퀘어 지하1층 B1079, 1080호 포36거리', '서울 마포구 월드컵북로 396 디지털파빌리온 지하1층 1041호', '서울 마포구 월드컵북로 396 누리꿈스퀘어 지하1층 1109-1호',]
+export const REPRESENTATIVE_FOOD = ['김치찌개', '태국음식(나시고랭, 팟타이 등)', '샤브샤브(칼국수)', '수제버거', '샤브샤브', '일식(규카츠, 덮밥)', '우육면, 탄탄면, 샤오롱바오', '우육면, 샤오롱바오', '곰탕', '돼지불백, 제주흑돼지 등', '차돌 쌀국수', '중식(짜장, 짬뽕 등)', '돈카츠(로스카츠, 히레카츠)', '가츠동', '나시고랭, 팟타이 등', '제주해장국', '경양식 돈까스, 함박스테이크 등', '샤브샤브', '캘리포니아롤', '일식 덮밥(사케동, 스테이크덮밥 등)', '양식(파스타, 빠네) - 데이트 추천', '평양냉면, 갈비탕(+고기집)', '꼬막 음식점', '참치 전문점(강사님 사주세요)', '보쌈정식, 감자옹심이 등', '양식(피자, 파스타 등) - 데이트 추천', '낙지볶음, 낙곱새볶음', '돈까스', '중식(짜장, 짬뽕 등)', '탄탄멘, 스테이크덮밥', '콩나물 국밥', '순대국밥, 나주곰탕 등', '국밥(a.k.a. 부산아지매)', '양평 해장국', '뼈해장국, 삼계탕 등', '햄버거', '햄버거', '햄버거', '구내식당','김밥천국 같은 곳', '김밥집', '김밥집', '김밥천국 같은 곳', '김밥 천국임.', '덮밥 류', '김밥집', '도시락', '쌀국수', '라면', '순두부찌개집(default 비빔밥 가능)', '소,머리국밥', '초밥', '일식류의 덮밥, 카레, 라멘', '돈까스', '쌀국수', '쌀국수', '떡볶이', ];
+export const PRICE_PRE_PERSON = ['10000', '20000u', '11000', '11700u', '14000', '19000','9500','15000', '12500', '10000', '15000', '10000', '15000','15000', '12000', '11000', '11000', '15000', '10000', '15000', '16000', '17000', '13000', '35000', '13000', '20000u', '14000', '12000', '10000', '10800', '9000', '13000', '11500', '15000', '11000', '7000', '7000', '7000', '6500', '8000', '김밥 한 줄 5000원, 닭강정 세트 10000', '한 줄 5500', '8000', '8000', '8500', '8500', '7000', '10000', '5000', '8500', '9000', '12900원 ~ 14900', '13000', '11900', '9900', '11000', '14000', ];
+export const MARKET_NAME = ['김치도가 상암점', '생어거스틴 상암DMC점', '등촌샤브칼국수 상암점', '바스버거 상암DMC점', '청록미나리식당 상암점', '후라토식당 상암직영점','룽탕 상하이딤섬', '샤오바오 상암점', '서당골 상암점', '동남집', '미분당 상암점', '홍콩반점0410 상암DMC점', '은옥', '온돈부리', '온타이키친', '도락', '경양식당', '채선당 상암DMC점', '미미롤집', '장정정 상암정', '달리181 상암점', '배꼽집 본점', '여수낭만식당', '참치공방 상암점', '시골보쌈&감자옹심이 상암점', '뚜띠쿠치나 상암점', '개미집 상암점', '마이클돈까스 본점', '용무있습니까 상암점', '미도인 상암엠비씨', '전주현대옥 상암DMC점', '김둘레순대국 상암2/4호점', '부산아지매국밥 상암DMC점', '상암 양평 해장국', '푸짐한왕뼈감자탕', '맥도날드 서울상암DMC', '롯데리아 상암점', 'KFC 상암DMC점', 'JTBC 구내식당', '최가네김밥', '오토김밥상암점', '청담마녀김밥 상암점', '김가네상암푸르지오', '김밥천국 디엠씨점', '오늘은덮밥', '여우애김밥 상암DMC점', '한솥도시락 상암DMC점', '포메인레드 상암DMC점', '틈새라면', '새벽집', '용길이네소머리국밥', '스시앤도시락 상암점', '핵밥 상암DMC점', '돈까스브로스 상암DMC점', '포36거리 상암', '동남쌀국 상암점', '응급실국물떡볶이 상암점', ];
+
+// 우리FISA 수업 반 학생들
+export const SERVICE_CALSS_MEMBERS = createNameList([
+    "윤영찬", "권민지", "권지윤", "김새봄", "김성엽", "김성준", "김소민", "김승호", "김윤권", "김지연",
+    "남승현", "노영재", "마서영", "박찬진", "배승혁", "서용준", "서채연", "신희원", "여은동", "유승한",
+    "윤선영", "윤예진", "윤태경", "이서연", "이소연", "이승준", "이원빈", "이정민", "이한비", "이현경",
+    "임지섭", "조윤주", "조현식", "차승훈", "허연규", "황유환", "황혜영", "구본훈"
+]);
+
+// GPT가 만들어준 가명들
+export const CLOUD_ENGINEERING_MEMBERS = createNameList([
+"박준형", "김민정", "이수민", "장현우", "최은영", "한지호", "강소라", "신동엽", "오현석", "이재영",
+"문서영", "황석현", "이상민", "배지훈", "조민수", "차유나", "고은비", "김도윤", "박혜진", "윤태훈",
+"정다연", "이한솔", "송지훈", "권나윤", "장동민", "김시현", "노정훈", "오지현", "백승환", "전민아"
+]);
+  
+// GPT가 만들어준 가명들
+export const AI_MEMBERS = createNameList([
+"김수현", "박서연", "이현준", "최하은", "한준혁", "정민지", "신유진", "문지호", "오준수", "강민경",
+"최정훈", "이유진", "김태환", "배민지", "조영우", "황보라", "고재훈", "서유나", "임태현", "박은지",
+"정하윤", "김정훈", "이민지", "박현우", "유다영", "장성민", "차예린", "윤지훈", "문채연", "오승준"
+]);
+
+sortNameList(SERVICE_CALSS_MEMBERS);
+sortNameList(CLOUD_ENGINEERING_MEMBERS);
+sortNameList(AI_MEMBERS);
