+++ js/util.js
@@ -1,0 +1,7 @@
+export function createNameList(names) {
+    return names.map((name) => ({
+      name,
+      excluded: false,
+    }));
+}
+
